+++ src/Audit.EntityFramework/CommandEvent.cs
@@ -5,7 +5,10 @@ using System.Data;
 
 namespace Audit.EntityFramework
 {
-    public class CommandEvent
+    /// <summary>
+    /// Event information for command interception
+    /// </summary>
+    public class CommandEvent : InterceptorEventBase
     {
         /// <summary>
         /// The command method (NonQuery, Scalar, Reader)
@@ -16,22 +19,6 @@ namespace Audit.EntityFramework
         /// </summary>
         public CommandType CommandType { get; set; }
         /// <summary>
-        /// The database name
-        /// </summary>
-        public string Database { get; set; }
-        /// <summary>
-        /// A unique identifier for the database connection.
-        /// This identifier is primarily intended as a correlation ID for logging and debugging such
-        /// that it is easy to identify that multiple events are using the same or different database connection.
-        /// </summary>
-        public string ConnectionId { get; set; }
-        /// <summary>
-        /// A unique identifier for the context instance and pool lease, if any.
-        /// This identifier is primarily intended as a correlation ID for logging and debugging such
-        /// that it is easy to identify that multiple events are using the same or different context instances.
-        /// </summary>
-        public string ContextId { get; set; }
-        /// <summary>
         /// The command text
         /// </summary>
         public string CommandText { get; set; }
@@ -39,18 +26,6 @@ namespace Audit.EntityFramework
         /// The parameter values
         /// </summary>
         public Dictionary<string, object> Parameters { get; set; }
-        /// <summary>
-        /// Boolean value indicating whether the call is asynchronous
-        /// </summary>
-        public bool IsAsync { get; set; }
-        /// <summary>
-        /// Boolean to indicate success. Null until command is executed
-        /// </summary>
-        public bool? Success { get; set; }
-        /// <summary>
-        /// The exception error message when Success is false
-        /// </summary>
-        public string ErrorMessage { get; set; }
         /// <summary>
         /// The result of the command execution
         /// </summary>